微修正

+++ rasp_test/controller/basic_controller.cpp
@@ -66,7 +66,7 @@ void basic_controller::update_arm(std::map<std::string, float>& normalized_contr
 
 	// * IB : Identifier Button (識別ボタン) : 腕の機能が割り当てられたボタン
 	// 1. IB						 : 状態を1ずつ遷移
-	// 2. index + IB + (+ or - or 0) : 指定番号の状態に遷移.さらにそこから微調整可能.
+	// 2. index + IB + (+ or - or nothing) : 指定番号の状態に遷移.さらにそこから微調整可能.
 	for (const auto& i : _arm_abilities_name) {
 		float ib_state = normalized_controller_state[config.key_config<std::string>("IB_" + i)];
 
@@ -79,7 +79,7 @@ void basic_controller::update_arm(std::map<std::string, float>& normalized_contr
 				_arm_adjusting_values[i] -= normalized_controller_state[config.key_config<std::string>("arm_adjusting_-")];
 			} else {
 				// 1. IB : 状態を1ずつ遷移
-				if (++_arm_abilities_position_index[i] >= config.setting<int>("arm_abilities_position_num")) {
+				if (++_arm_abilities_position_index[i] >= config.setting<size_t>("arm_abilities_position_num")) {
 					_arm_abilities_position_index[i] = 0;
 				}
 			}
